Better CI-comment finding
If someone's posted a real comment on a review after a Jenkins failure
(e.g. actual review comments, not bot spew), then the CI comment will
not be in the last 2 comments any more. Now we go look at all the
comments for the newest patch set (or as best we can determine it) to
find the CI and ER comments.